backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
chardet==5.2.0

# Note: 推荐使用 Python 3.11 或 3.12
# Python 3.14 可能存在依赖兼容性问题
